(SIMP-4482) Pre-popluate pki_files dirs
Pre-populate the /var/simp/environments/simp/site_files/pki_files/
directory tree and set its ownership and permissions appropriately.
This is part of the fix to the failure of SIMP to bootstrap on
a system on which root's umask has already been restricted to 077.

SIMP-4482 #comment simp-environment pki_files fix